+++ WatchFriends/app/src/main/java/nmct/jaspernielsmichielhein/watchfriends/helper/Contract.java
@@ -7,5 +7,6 @@ public class Contract {
     public static final String MOVIEDB_IMAGE_BASE_URL = "https://image.tmdb.org/t/p/w342/";
 
     //WATCHFRIENDS
-    public static final String WATCHFRIENDS_BASE_URL = "http://172.30.248.101:8080/data/";
+    //TODO: url aanpassen!
+    public static final String WATCHFRIENDS_BASE_URL = "https://raw.githubusercontent.com/WatchFriends/Backend/master/data/";
 }
