+++ WatchFriends/app/src/main/java/nmct/jaspernielsmichielhein/watchfriends/viewmodel/SeasonFragmentViewModel.java
@@ -9,6 +9,7 @@ import android.view.View;
 import nmct.jaspernielsmichielhein.watchfriends.BR;
 import nmct.jaspernielsmichielhein.watchfriends.databinding.FragmentSeasonBinding;
 import nmct.jaspernielsmichielhein.watchfriends.helper.ApiHelper;
+import nmct.jaspernielsmichielhein.watchfriends.helper.ApiMovieDbHelper;
 import nmct.jaspernielsmichielhein.watchfriends.helper.Interfaces;
 import nmct.jaspernielsmichielhein.watchfriends.model.Season;
 import rx.functions.Action1;
@@ -48,7 +49,7 @@ public class SeasonFragmentViewModel extends BaseObservable {
 
     public void loadSeason() {
         final SeasonFragmentViewModel that = this;
-        ApiHelper.subscribe(ApiHelper.getMoviedbServiceInstance().getSeason(seriesId, seasonNumber),
+        ApiHelper.subscribe(ApiMovieDbHelper.getMoviedbServiceInstance().getSeason(seriesId, seasonNumber),
             new Action1<Season>() {
                 @Override
                 public void call(Season season) {
