+++ WatchFriends/app/src/main/res/layout/fragment_home.xml
@@ -33,7 +33,12 @@
 
             <android.support.v7.widget.RecyclerView
                 android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
+                android:layout_height="match_parent"
+                android:layout_margin="0dp"
+                android:orientation="vertical"
+                android:id="@+id/rvLists"
+                app:items="@{viewmodel.items}"/>
+
         </LinearLayout>
     </ScrollView>
 </layout>
